[feat] Add/Edit/Delete Markers
Defined routes to match Adding/Editing/Deleting for Markers given a user is logged in. Wrote functionality to add newly constructed markers to db. Added test cases to check that 1) Adding a marker causes it to appear in the Members Internal page. 2) Editing a marker cause its value to change accordingly.